pyproject.toml
[project]
name = "mfb-webapp-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "teleporterx"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "motor (>=3.6.0,<4.0.0)",
    "pydantic (>=2.10.5,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==3.2.2)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
